requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
pymongo==4.6.0
psycopg2-binary==2.9.9
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.12.2
python-multipart==0.0.6
openpyxl==3.1.2
pytest==7.4.3